+++ musestat/config.py
@@ -2,7 +2,7 @@
 Configuration constants for MuseStat.
 """
 
-__version__ = "1.1.0"
+__version__ = "1.2.0"
 __author__ = "Tim Gatzke"
 __email__ = "post@tim-gatzke.de"
 __website__ = "https://tim-gatzke.de"