+++ musestat/ui/__init__.py
@@ -3,6 +3,7 @@
 from .display import display_statistics, print_minimalist, list_manuscript_files
 from .panels import *
 from .tables import *
+from .visualizations import *
 
 __all__ = [
     'display_statistics',